Complete Lab 1 - Week 2 Exercises
Completed on 4/8/24.

Exercises complete:
5.1 - Store the following table as an array of objects, using the first row of the table as keys for
the objects

Exercise 6.1. Using the array of objects from Exercise 5.1, create a new array that contains only the
Item names but sorted according to quantity, with the item associated with the highest quantity included
first.

Exercise 6.2. Similar to Exercise 6.1, using the array of objects from Exercise 5.1, create a new array
that contains only the Item names sorted in alphabetical order.
Hint: Start a good habit and use the localeCompare function to compare strings.

Exercise 7.1. Given an array of numbers: [5, 0.3, NaN, 2, -4, 2.1, Infinity], sort the array so
that the numbers are listed from smallest to largest with the special number Infinity at the end and NaN
at the beginning (before the other numbers).
Hint: Use the special number function defined in Table 2 to test for NaN

Exercise 8.1. Set up a for loop that iterates through the array in Exercise 7.1 and prints the number to
the console only if it’s finite. Format the numbers to have 2 decimals.